pyproject.toml
[project]
name = "wizard-wizard"
version = "0.1.0"
description = "An AI Wizard for building a DnD Wizard"
readme = "README.md"
requires-python = ">=3.13.9"
dependencies = [
    "async-lru>=2.0.5",
    "google-adk>=1.19.0",
    "litellm>=1.80.5",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "rapidfuzz>=3.14.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.6",
    "ty>=0.0.1a27",
]
